+++ src/hipay_enterprise/classes/helper/HipayNotification.php
@@ -167,7 +167,7 @@ class HipayNotification
 
         $notifications = $this->dbMaintenance->getWaitingNotificationsAndUpdateStatus(
             NotificationStatus::IN_PROGRESS,
-            Configuration::get('HIPAY_NOTIFICATION_THRESHOLD')
+            $this->configHipay['account']['global']['notification_max_retry'] ?? 50
         );
 
         $totalError = 0;
@@ -823,7 +823,7 @@ class HipayNotification
 
             OrderSlip::create(
                 $order,
-                [$product],
+                [$product]
             );
         }